+++ src/test/java/sagex/phoenix/metadata/search/RegexTVFilenameScraperTest.java
@@ -53,6 +53,12 @@ public class RegexTVFilenameScraperTest {
         testTV(scraper, "THE_WALKING_DEAD-S02E11.mkv", "THE WALKING DEAD", "02","11", null,null,null);
         testTV(scraper, "THE.WALKING.DEAD.S02E11.mkv", "THE WALKING DEAD", "02","11", null,null,null);
 
+        testTV(scraper, "THE WALKING DEAD S02E01-06/BDMV/", "THE WALKING DEAD", "02", "01", "06", null,null);
+
+
+        // we don't autoexpand camel case names for normal files.
+        testTV(scraper, "SignedSealedDelivered_S02E11_TimeToStartLivin.mkv", "SignedSealedDelivered", "02", "11", null, null, null);
+
         testAiring(scraper,"The100-S03E07-Thirteen-12775038-0.mpg","The 100", "03","07","Thirteen","12775038");
 
         testNoMatch(scraper, "Savages.2012.720p.bluray.x264-sparks.mkv");