+++ src/test/java/sagex/phoenix/metadata/search/SearchQueryFactoryTest.java
@@ -6,6 +6,7 @@ import sagex.phoenix.Phoenix;
 import sagex.phoenix.metadata.MediaType;
 import sagex.phoenix.util.Hints;
 import sagex.phoenix.vfs.IMediaFile;
+import sagex.phoenix.vfs.MediaResourceType;
 import sagex.phoenix.vfs.impl.FileMediaFile;
 import sagex.phoenix.vfs.impl.FileResourceFactory;
 import test.InitPhoenix;
@@ -35,4 +36,20 @@ public class SearchQueryFactoryTest {
         assertEquals("Pokémon Go", q.getRawTitle());
     }
 
+    @Test
+    public void createQueryFromBluray() throws Exception {
+        //  \\SERVER-W1\MOVIES3\Doctor Strange\BDMV
+
+        File mfile = FilesTestCase.makeFile("test/Doctor Strange/BDMV/video.m2ts");
+        IMediaFile mf = (IMediaFile) FileResourceFactory.createResource(mfile);
+
+        // just verify that it thinks we have a bluray file
+        assertTrue(mf.isType(MediaResourceType.BLURAY.value()));
+        assertEquals("Doctor Strange", mf.getTitle());
+
+        SearchQuery q = Phoenix.getInstance().getSearchQueryFactory().createQueryFromFilename(mf, MediaType.MOVIE, new Hints());
+        System.out.println(q);
+        assertNotNull("Search Querries should never be null!!", q);
+        assertEquals("Doctor Strange", q.getRawTitle());
+    }
 }